Make error Toast a nicer Snackbar, refresh theme on change

+++ app/src/main/java/fr/outadoc/quickhass/preferences/PreferenceRepository.kt
@@ -5,7 +5,7 @@ interface PreferenceRepository {
     var instanceBaseUrl: String
     val altInstanceBaseUrl: String?
     var accessToken: String
-    val refreshIntervalSeconds: Int
+    val refreshIntervalSeconds: Long
     val showWhenLocked: Boolean
     val theme: String
